Set reset_after on the first increment
This commit changes the behaviour of the reset_after variable.
Previously this variable would only be set when the limit was exceeded.
The variable is now set on the first increment of a new cycle.
(A cycle starting at instance creation or counter reset)

This is more in line with how Xbox Live measures requests for rate limiting.

* Set reset_after on the first increment
* Create __set_reset_after function
* Add code comments